+++ src/Simpletools/Db/Cassandra/Type/BigInt.php
@@ -2,13 +2,18 @@
 
 namespace Simpletools\Db\Cassandra\Type;
 
-class BigInt
+class BigInt implements \JsonSerializable
 {
     protected $_value;
 
     public function __construct($bigint)
     {
-        $this->_value = new \Cassandra\BigInt((string) $bigint);
+			if(is_numeric($bigint))
+				$this->_value = new \Cassandra\BigInt((string) $bigint);
+			elseif($bigint instanceof \Cassandra\BigInt)
+				$this->_value = $bigint;
+			else
+				throw new \Exception("BigInt: Value is not numeric");
     }
 
     public function value()
@@ -16,10 +21,10 @@ class BigInt
         return $this->_value;
     }
 
-//    public function value()
-//    {
-//        return (string) $this->_value;
-//    }
+		public function jsonSerialize()
+		{
+			return (string)$this->value();
+		}
 
     public function __toString()
     {
@@ -30,4 +35,10 @@ class BigInt
     {
         return (int) (string) $this->value();
     }
+
+    public function setDefault()
+		{
+			$this->_value = new \Cassandra\BigInt('0');
+			return $this;
+		}
 }