+++ src/Simpletools/Db/Cassandra/Type/Timeuuid.php
@@ -8,23 +8,33 @@ class Timeuuid implements \JsonSerializable
 {
     protected $_value;
 
-    public function __construct($timeSec=null)
+    public function __construct($time=null)
     {
-        if(is_numeric($timeSec)) {
+			if(is_numeric($time) && strlen($time) <= 10)
+				$this->_value = new \Cassandra\Timeuuid((int)$time);
+			if(is_string($time) && strlen($time) == 36)
+				$this->_value = new \Cassandra\Timeuuid($time);
+			elseif (is_string($time) && ($time = strtotime($time)))
+				$this->_value = new \Cassandra\Timeuuid($time);
+			elseif ($time===null)
+				$this->_value = new \Cassandra\Timeuuid();
+			elseif ($time instanceof \Cassandra\Timeuuid)
+				$this->_value = $time;
+			else
+				throw new \Exception("Timeuuid: Value is not a timestamp or date");
 
-            echo (int) $timeSec;
-            $this->_value = new \Cassandra\Timeuuid((int)$timeSec);
-        }
-        elseif($timeSec!==null)
-            throw new Exception("Wrong argument type",401);
-        else
-            $this->_value = new \Cassandra\Timeuuid();
+
     }
 
     public function value()
     {
-        return ($this->_value = (string) $this->_value);
+        return $this->_value;
     }
+
+		public function toInt()
+		{
+			return (int) $this->value();
+		}
 
     public function jsonSerialize()
     {
@@ -33,8 +43,14 @@ class Timeuuid implements \JsonSerializable
 
     public function __toString()
     {
-        return $this->value();
+        return (string)$this->value();
     }
+
+	public function setDefault()
+	{
+		$this->_value = new \Cassandra\Timeuuid(0);
+		return $this;
+	}
 
     public function time()
     {